@@ -46,6 +46,7 @@
     "@salesforce/command": "^5.1.2",
     "@salesforce/core": "^3.19.0",
     "@salesforce/ts-types": "^1.5.20",
+    "fast-xml-parser": "^4.0.7",
     "tslib": "^2"
   },
   "devDependencies": {
@@ -53,16 +54,16 @@
     "@salesforce/dev-config": "^3.0.1",
     "@salesforce/dev-scripts": "^2.0.2",
     "@salesforce/plugin-command-reference": "^1.3.17",
+    "@salesforce/prettier-config": "^0.0.2",
     "@salesforce/ts-sinon": "^1.3.21",
     "@semantic-release/git": "^9.0.0",
-    "@types/chai": "^4.2.11",
-    "@types/mocha": "^7.0.2",
-    "@types/node": "^14.0.9",
     "@types/xml2js": "^0.4.5",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/parser": "^4.2.0",
     "bluebird": "3.7.2",
     "chai": "^4.2.0",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^7",
+    "eslint": "^7.27.0",
     "eslint-config-prettier": "^6.15.0",
     "eslint-config-salesforce": "^0.1.6",
     "eslint-config-salesforce-license": "^0.1.6",
@@ -72,20 +73,20 @@
     "eslint-plugin-jsdoc": "^39.3.2",
     "eslint-plugin-prettier": "^3.4.1",
     "globby": "^11.0.1",
-    "husky": "^7",
+    "husky": "^7.0.4",
     "mocha": "^9.1.3",
     "nyc": "^15.1.0",
     "oclif": "^3.0.1",
     "prettier": "^2.6.2",
+    "pretty-quick": "^3.1.0",
     "semantic-release": "^17.0.8",
     "shx": "0.2.2",
-    "sinon": "9.0.2",
-    "ts-node": "^8.10.2",
-    "typescript": "^4.3.2",
-    "xml2js": "^0.4.23"
+    "sinon": "10.0.0",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.2"
   },
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=14.0.0"
   },
   "files": [
     "/csvtojson",
@@ -124,17 +125,24 @@
   },
   "repository": "github:salesforcecli/plugin-custom-metadata",
   "scripts": {
+    "build": "sf-build",
+    "clean": "sf-clean",
+    "clean-all": "sf-clean all",
     "clean:lib": "shx rm -rf lib && shx rm -rf coverage && shx rm -rf .nyc_output && shx rm -f oclif.manifest.json",
-    "compile": "tsc -b",
-    "build": "yarn run clean:lib && yarn compile",
-    "postpack": "rm -f oclif.manifest.json",
-    "posttest": "yarn test:deprecation-policy && yarn test:command-reference",
-    "prepack": "rm -rf lib && tsc && oclif-dev manifest && oclif-dev readme",
-    "test": "nyc mocha --forbid-only \"test/**/*.test.ts\"",
-    "version": "oclif-dev readme && git add README.md",
-    "release": "semantic-release -p @semantic-release/git -p @semantic-release/npm -p @semantic-release/github -p @semantic-release/commit-analyzer",
-    "test:deprecation-policy": "./bin/run snapshot:compare",
-    "test:command-reference": "./bin/run commandreference:generate -p @salesforce/plugin-custom-metadata --erroronwarnings"
+    "compile": "sf-compile",
+    "docs": "sf-docs",
+    "format": "sf-format",
+    "lint": "sf-lint",
+    "postpack": "shx rm -f oclif.manifest.json",
+    "posttest": "yarn lint && yarn test:deprecation-policy && yarn test:json-schema",
+    "prepack": "sf-prepack",
+    "prepare": "sf-install",
+    "pretest": "sf-compile-test",
+    "test": "sf-test",
+    "test:command-reference": "./bin/dev commandreference:generate --erroronwarnings",
+    "test:deprecation-policy": "./bin/dev snapshot:compare",
+    "test:json-schema": "./bin/dev schema:compare",
+    "version": "oclif readme"
   },
   "publishConfig": {
     "access": "public"
