@@ -4,8 +4,6 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-
-export interface ErrorMessage {
-  errorCode: string;
-  errorMessage: string;
-}
+module.exports = {
+  extends: ['eslint-config-salesforce-typescript', 'eslint-config-salesforce-license'],
+};
