@@ -5,8 +5,8 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { CustomField } from './customField';
-
+import { Record } from 'jsforce';
+import { CustomField } from 'jsforce/api/metadata';
 export interface CreateConfig {
   typename: string;
   recordname: string;
@@ -14,7 +14,7 @@ export interface CreateConfig {
   inputdir: string;
   outputdir: string;
   protected?: boolean;
-  varargs?: object;
+  varargs?: Record;
   fileData?: CustomField[];
   ignorefields?: boolean;
 }
